Change file structure and add doc
