+++ kandy/src/main/java/pt/hugofernandes/kandy/flow/Lifecycle.kt
@@ -4,13 +4,18 @@ import androidx.lifecycle.*
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
 
+/** A named reference to this [LifecycleOwner]'s instance. */
+@PublishedApi
+internal inline val LifecycleOwner.lifecycleOwner: LifecycleOwner
+    get() = this
+
 /**
  * Accepts the [collector] on the provided [flow] and emits values into it in a lifecycle-aware
  * manner. The collect operation is launched using the [owner]'s [LifecycleCoroutineScope] and
  * automatically repeated on lifecycle event [Lifecycle.State.STARTED].
  * @param owner The [LifecycleOwner] which the lifecycle-aware [Flow.collect] operation is tied to.
  * @param flow The data stream from which values will be emitted.
- * @param coroutineDispatcher The [CoroutineDispatcher] used to launch the collect operations.
+ * @param coroutineDispatcher The [CoroutineDispatcher] used to launch the collect operation.
  * @param collector The consumer of emitted data.
  */
 @PublishedApi
@@ -20,9 +25,7 @@ internal inline fun <T> lifecycleAwareCollect(
     coroutineDispatcher: CoroutineDispatcher,
     crossinline collector: suspend CoroutineScope.(T) -> Unit
 ) = owner.lifecycleScope.launch(coroutineDispatcher) {
-    owner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-        flow.collect { collector(it) }
-    }
+    owner.repeatOnLifecycle(Lifecycle.State.STARTED) { flow.collect { collector(it) } }
 }
 
 /**
