+++ kandy/src/main/java/pt/hugofernandes/kandy/ktx/LiveDataEx.kt
@@ -1,11 +1,11 @@
 package pt.hugofernandes.kandy.ktx
 
-import android.os.Handler
-import android.os.Looper
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MediatorLiveData
-import androidx.lifecycle.Observer
+import android.os.*
+import android.view.View
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.*
+import pt.hugofernandes.kandy.flow.lifecycleOwner
 
 /**
  * Debounce updates on the given [LiveData].
@@ -79,3 +79,37 @@ fun <T> LiveData<T>.observe(owner: LifecycleOwner, observer: (t: T?) -> Unit) {
 }
 
 class NonNullMediatorLiveData<T> : MediatorLiveData<T>()
+
+//region Lifecycle-aware observe
+/**
+ * Adds the given [onChange] block as this [LiveData]'s observer within the this `Activity` lifespan
+ * and returns a reference to the observer.
+ */
+context(AppCompatActivity)
+inline fun <T> LiveData<T>.observe(crossinline onChange: (T) -> Unit) =
+    observe(lifecycleOwner, onChange)
+
+/**
+ * Adds the given [onChange] block as this [LiveData]'s observer within the this [Fragment]'s
+ * `viewLifecycleOwner` lifespan and returns a reference to the observer.
+ */
+context(Fragment)
+inline fun <T> LiveData<T>.observe(crossinline onChange: (T) -> Unit) =
+    observe(viewLifecycleOwner, onChange)
+
+/**
+ * Adds the given [onChange] block as this [LiveData]'s observer within the this `Service` lifespan
+ * and returns a reference to the observer.
+ */
+context(LifecycleService)
+inline fun <T> LiveData<T>.observe(crossinline onChange: (T) -> Unit) =
+    observe(lifecycleOwner, onChange)
+
+/**
+ * Adds the given [onChange] block as this [LiveData]'s observer within the this [View]'s lifespan
+ * and returns a reference to the observer.
+ */
+context(View, LifecycleOwner)
+inline fun <T> LiveData<T>.observe(crossinline onChange: (T) -> Unit) =
+    observe(lifecycleOwner, onChange)
+//endregion
